pyproject.toml
[tool.poetry]
name = "update-poms"
version = "0.1.0"
description = ""
authors = ["Ian <ian_bacher@brown.edu>"]
license = "MPL2"

[tool.poetry.dependencies]
python = "^3.9"
GitPython = "^3.1.15"
"github3.py" = "^2.0.0"
lxml = "^4.6.3"

[tool.poetry.dev-dependencies]
black = "^21.5b0"
flake8 = "^3.9.1"
mypy = "^0.812"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
